refactor(dashboard): extract ArtistStats presentational component
Move the artist stats computation + badge cards out of the 650-line
dashboard/artist.vue monolith into Card/Dashboard/ArtistStats.vue (props in,
no emits/v-model — zero reactivity risk). First slice of the dashboard
decomposition.
